Add working conn_mq + refactoring + bad instance

+++ kozlov.ilya/lab2/goat/src/goat.cpp
@@ -5,8 +5,7 @@
 #include <csignal>
 #include <memory.h>
 #include <random>
-
-int Goat::last_err = 0;
+#include <cstring>
 
 Goat& Goat::GetInstance(int host_pid)
 {
@@ -14,29 +13,47 @@ Goat& Goat::GetInstance(int host_pid)
   return inst;
 }
 
-Goat::Goat(int host_pid) : connection(host_pid, false)
+Goat::Goat(int host_pid)
 {
   std::cout << "host Wolf pid is: " << host_pid << std::endl;
   this->host_pid = host_pid;
   signal(SIGTERM, SignalHandler);
-  semaphore = sem_open(SEM_NAME, 0);
-  if (semaphore == SEM_FAILED)
-  {
-    std::cout << "ERROR: sem_open failed with error = " << errno << std::endl;
-    last_err = errno;
-  }
-  else
-  {
-    std::cout << "pid of created Goat is: " << getpid() << std::endl;
-    kill(host_pid, SIGUSR1);
-  }
 }
 
-Goat::~Goat()
+bool Goat::OpenConnection()
 {
-  std::cout << "~Goat()" << std::endl;
-  sem_post(semaphore);
-  sem_close(semaphore);
+  bool res = false;
+  if (connection.Open(host_pid, false))
+  {
+    semaphore = sem_open(SEM_NAME, 0);
+    if (semaphore == SEM_FAILED)
+    {
+      std::cout << "ERROR: sem_open failed with error = " << strerror(errno) << std::endl;
+    }
+    else
+    {
+      res = true;
+      std::cout << "pid of created Goat is: " << getpid() << std::endl;
+      kill(host_pid, SIGUSR1);
+    }
+  }
+  return res;
+}
+
+void Goat::Terminate(int signum)
+{
+  kill(host_pid, SIGUSR2);
+  std::cout << "Goat::Terminate()" << std::endl;
+  if (errno != 0)
+  {
+    std::cout << "Failing with error = " << strerror(errno) << std::endl;
+  }
+  if (connection.Close() && sem_post(semaphore) == 0 && sem_close(semaphore) == 0)
+  {
+    exit(signum);
+  }
+  std::cout << "Terminating error: " << strerror(errno) << std::endl;
+  exit(errno);
 }
 
 #pragma clang diagnostic push
@@ -46,17 +63,29 @@ void Goat::Start()
   struct timespec ts;
   ts.tv_sec = 5;
   ts.tv_nsec = 0;
+  int skipped_msgs = 0;
   while (true)
   {
-    sleep(1);
+    //sleep(1);
     if (sem_timedwait(semaphore, &ts) == -1)
     {
-      kill(host_pid, SIGUSR1);
-      exit(errno);
+      Terminate(errno);
     }
     Memory msg;
     if (connection.Read(&msg, 0))
     {
+      if (msg.owner == GOAT)
+      {
+        skipped_msgs++;
+        std::cout << "Host wrote nothing " << skipped_msgs << " times, skipping..." << std::endl;
+        if (skipped_msgs >= MAX_SKIPPED_MSGS)
+        {
+          Terminate(SIGTERM);
+        }
+        sem_post(semaphore);
+        continue;
+      }
+      skipped_msgs = 0;
       std::cout << "Wolf number: " << msg.number << std::endl;
       std::cout << "Status: " << ((msg.status == ALIVE) ? "alive" : "dead") << std::endl;
       if (msg.status == ALIVE)
@@ -68,6 +97,7 @@ void Goat::Start()
         msg.number = GetRand(50);
       }
       std::cout << "Goat number: " << msg.number << std::endl;
+      msg.owner = GOAT;
       connection.Write((void *)&msg, sizeof(msg));
     }
     sem_post(semaphore);
@@ -85,8 +115,6 @@ int Goat::GetRand(int right)
 
 void Goat::SignalHandler(int signum)
 {
-  if (signum == SIGTERM)
-  {
-    exit(SIGTERM);
-  }
+  Goat& instance = Goat::GetInstance(30);
+  instance.Terminate(signum);
 }
